+++ xenon-common/src/main/java/com/vmware/xenon/common/http/netty/NettyHttpListener.java
@@ -112,7 +112,12 @@ public class NettyHttpListener implements ServiceRequestListener {
 
     @Override
     public void setSSLContextFiles(URI certFile, URI keyFile) throws Throwable {
-        this.sslContext = SslContext.newServerContext(new File(certFile), new File(keyFile));
+        this.setSSLContextFiles(certFile, keyFile, null);
+    }
+
+    @Override
+    public void setSSLContextFiles(URI certFile, URI keyFile, String keyPassphrase) throws Throwable {
+        this.sslContext = SslContext.newServerContext(new File(certFile), new File(keyFile), keyPassphrase);
     }
 
     @Override